neuraplay/requirements.txt
Django>=5.0
gunicorn
python-dotenv
google-genai
pytest 
pytest-django 
djangorestframework
firebase-admin
djangorestframework-simplejwt
google-cloud-texttospeech 
pyrebase4


